test: move FileHandle postMessage tests to a dedicated file
The full worker_threads.test.ts has pre-existing 5-second subprocess
timeouts under debug+ASAN in the constrained gate environment. Keep the
new FileHandle-transfer tests in their own file so the gate's before/after
comparison is not drowned out by that noise. Also extend the in-use and
rollback tests so they transfer successfully afterwards, which makes each
test in the file fail without the fix.
